+++ src/components/Link.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { Link } from 'react-router-dom';
-import { colors, fonts } from '../../styles/style_guide';
+import { colors, fonts } from '../styles/style_guide';
 
 
 export default styled(Link)`
